MBX-2350 Small refactoring
