+++ Mindbox/InAppMessages/InAppMapper/API/InAppConfigResponse.swift
