+++ Mindbox/InAppMessages/InAppMapper/API/InAppConfigurationAPI.swift
