+++ Mindbox/InAppMessages/InAppMapper/API/TargetingModel.swift
