+++ Mindbox/InAppMessages/InAppMapper/API/Targetings/AndTargeting.swift
