+++ Mindbox/InAppMessages/InAppMapper/API/Targetings/CategoryIDInTargeting.swift
