+++ Mindbox/InAppMessages/InAppMapper/API/Targetings/CategoryIDTargeting.swift
