+++ Mindbox/InAppMessages/InAppMapper/API/Targetings/CustomOperationTargeting.swift
