+++ Mindbox/InAppMessages/InAppMapper/API/Targetings/ProductSegmentTargeting.swift
