+++ Mindbox/InAppMessages/InAppMapper/API/Targetings/SegmentTargeting.swift
