+++ Mindbox/InAppMessages/InAppMapper/InAppConfig.swift
